[svn-r19897] Extend packed bits and testing to 64 bit integers.
Tested: local Linux

+++ tools/lib/h5tools.c
@@ -45,7 +45,7 @@ int         region_output; /* region output */
 #ifdef H5_HAVE_H5DUMP_PACKED_BITS
 int         packed_bits_num; /* number of packed bits to display */
 int         packed_data_offset; /* offset of packed bits to display */
-unsigned int packed_data_mask;  /* mask in which packed bits to display */
+unsigned long long packed_data_mask;  /* mask in which packed bits to display */
 #endif
 
 static h5tool_format_t h5tools_dataformat = {
